Renaming sim.h and sim.c to emulatorapi.h and emulatorapi.c respectively.
