+++ lib/emulatorapi.c
@@ -1,4 +1,4 @@
-#include "sim.h"
+#include "emulatorapi.h"
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
