+++ lib/emulatorapi.h
